[2.0] Refactor `TrackerAnnounceAlert` to use `event_t` enum and remove `SWIGTYPE_p_event_t`
 - Replaced `SWIGTYPE_p_event_t` with `event_t` enum from `tracker_manager.hpp`.
 - Modified `getEvent()` method in `TrackerAnnounceAlert` to use `event_t` and return its corresponding value.
 - Updated JNI bindings to correctly handle `event_t` in tracker announce alerts.
 - Removed redundant SWIGTYPE_p_event_t class and its references.
 - Added support for `event_t` in the SWIG interface files.

+++ src/main/java/com/frostwire/jlibtorrent/swig/libtorrent_jni.java
@@ -2401,7 +2401,7 @@ public class libtorrent_jni {
   public final static native String tracker_announce_alert_what(long jarg1, tracker_announce_alert jarg1_);
   public final static native long tracker_announce_alert_static_category_get();
   public final static native String tracker_announce_alert_message(long jarg1, tracker_announce_alert jarg1_);
-  public final static native long tracker_announce_alert_event_get(long jarg1, tracker_announce_alert jarg1_);
+  public final static native int tracker_announce_alert_event_get(long jarg1, tracker_announce_alert jarg1_);
   public final static native void tracker_announce_alert_version_set(long jarg1, tracker_announce_alert jarg1_, int jarg2);
   public final static native int tracker_announce_alert_version_get(long jarg1, tracker_announce_alert jarg1_);
   public final static native void delete_tracker_announce_alert(long jarg1);
